feat(MediaChannels): Support Media Channels (#1486)
