+++ esm.mjs
@@ -33,6 +33,7 @@ export const {
   GuildTemplate,
   Interaction,
   Invite,
+  MediaChannel,
   Member,
   Message,
   ModalSubmitInteraction,
