+++ lib/Client.js
@@ -584,18 +584,18 @@ class Client extends EventEmitter {
      * @arg {String} name The name of the channel
      * @arg {Number} [type=0] The type of the channel, either 0 (text), 2 (voice), 4 (category), 5 (news), 13 (stage), or 15 (forum)
      * @arg {Object | String} [options] The properties the channel should have. If `options` is a string, it will be treated as `options.parentID` (see below). Passing a string is deprecated and will not be supported in future versions.
-     * @arg {Array<Object>} [options.availableTags] The available tags that can be applied to threads in a forum channel. See [the official Discord API documentation entry](https://discord.com/developers/docs/resources/channel#forum-tag-object) for object structure (forum channels only, max 20)
+     * @arg {Array<Object>} [options.availableTags] The available tags that can be applied to threads in a forum/media channel. See [the official Discord API documentation entry](https://discord.com/developers/docs/resources/channel#forum-tag-object) for object structure (forum/media channels only, max 20)
      * @arg {Number} [options.bitrate] The bitrate of the channel (voice channels only)
-     * @arg {Number} [options.defaultAutoArchiveDuration] The default duration of newly created threads in minutes to automatically archive the thread after inactivity (60, 1440, 4320, 10080) (text/news/forum channels only)
+     * @arg {Number} [options.defaultAutoArchiveDuration] The default duration of newly created threads in minutes to automatically archive the thread after inactivity (60, 1440, 4320, 10080) (text/news/forum/media channels only)
      * @arg {Number} [options.defaultForumLayout] The default forum layout type used to display posts in forum channels (forum channels only)
-     * @arg {Object} [options.defaultReactionEmoji] The emoji to show in the add reaction button on a thread in a forum channel (forum channels only)
-     * @arg {Number} [options.defaultSortOrder] The default sort order type used to order posts in forum channels (forum channels only)
-     * @arg {Number} [options.defaultThreadRateLimitPerUser] The initial rateLimitPerUser to set on newly created threads in a channel (text/forum channels only)
+     * @arg {Object} [options.defaultReactionEmoji] The emoji to show in the add reaction button on a thread in a forum/media channel (forum/media channels only)
+     * @arg {Number} [options.defaultSortOrder] The default sort order type used to order posts in forum/media channels (forum/media channels only)
+     * @arg {Number} [options.defaultThreadRateLimitPerUser] The initial rateLimitPerUser to set on newly created threads in a channel (text/forum/media channels only)
      * @arg {Boolean} [options.nsfw] The nsfw status of the channel
      * @arg {String?} [options.parentID] The ID of the parent category channel for this channel
      * @arg {Array<Object>} [options.permissionOverwrites] An array containing permission overwrite objects
      * @arg {Number} [options.position] The sorting position of the channel
-     * @arg {Number} [options.rateLimitPerUser] The time in seconds a user has to wait before sending another message (0-21600) (does not affect bots or users with manageMessages/manageChannel permissions) (text/voice/stage/forum channels only)
+     * @arg {Number} [options.rateLimitPerUser] The time in seconds a user has to wait before sending another message (0-21600) (does not affect bots or users with manageMessages/manageChannel permissions) (text/voice/stage/forum/media channels only)
      * @arg {String} [options.reason] The reason to be displayed in audit logs
      * @arg {String} [options.topic] The topic of the channel (text channels only)
      * @arg {Number} [options.userLimit] The channel user limit (voice channels only)
@@ -1075,10 +1075,10 @@ class Client extends EventEmitter {
      * Create a thread in a channel
      * @arg {String} channelID The ID of the channel
      * @arg {Object} options The thread options
-     * @arg {Array<String>} [options.appliedTags] The IDs of the set of tags that have been applied to a thread in a forum channel (threads created in forum channels only, max 5)
+     * @arg {Array<String>} [options.appliedTags] The IDs of the set of tags that have been applied to a thread in a forum/media channel (threads created in forum/media channels only, max 5)
      * @arg {Number} [options.autoArchiveDuration] The duration in minutes to automatically archive the thread after recent activity, either 60, 1440, 4320 or 10080
      * @arg {Boolean} [options.invitable] Whether non-moderators can add other non-moderators to the thread (private threads only)
-     * @arg {Object} [options.message] The message to create with the thread (only use when creating a thread inside of a `GUILD_FORUM` channel). Note: When creating a forum channel thread, you must provide at least one of `content`, `embeds`, `stickerIDs`, `components`, or `files`
+     * @arg {Object} [options.message] The message to create with the thread (only use when creating a thread inside of a thread-only channel). Note: When creating a forum/media channel thread, you must provide at least one of `content`, `embeds`, `stickerIDs`, `components`, or `files`
      * @arg {Object} [options.message.allowedMentions] A list of mentions to allow (overrides default)
      * @arg {Boolean} [options.message.allowedMentions.everyone] Whether or not to allow @everyone/@here
      * @arg {Boolean} [options.message.allowedMentions.repliedUser] Whether or not to mention the author of the message being replied to
@@ -1111,8 +1111,8 @@ class Client extends EventEmitter {
      * @arg {String} options.name The thread channel name
      * @arg {Number} [options.rateLimitPerUser] The time in seconds a user has to wait before sending another message (0-21600) (does not affect bots or users with manageMessages/manageChannel permissions)
      * @arg {String} [options.reason] The reason to be displayed in audit logs
-     * @arg {Number} [options.type] The channel type of the thread to create. Either `10` (announcement thread, announcement channels only), `11` (public thread) or `12` (private thread). Note: Not required when creating a thread inside of a forum channel, it will always be public
-     * @arg {Object | Array<Object>} [file] A file object (or an Array of them). Only use when creating a thread inside of a forum channel
+     * @arg {Number} [options.type] The channel type of the thread to create. Either `10` (announcement thread, announcement channels only), `11` (public thread) or `12` (private thread). Note: Not required when creating a thread inside of a forum/media channel, it will always be public
+     * @arg {Object | Array<Object>} [file] A file object (or an Array of them). Only use when creating a thread inside of a forum/media channel
      * @arg {Buffer} file.file A buffer containing file data
      * @arg {String} file.name What to name the file
      * @returns {Promise<NewsThreadChannel | PublicThreadChannel | PrivateThreadChannel>}
@@ -1536,17 +1536,17 @@ class Client extends EventEmitter {
      * Edit a channel's properties
      * @arg {String} channelID The ID of the channel
      * @arg {Object} options The properties to edit
-     * @arg {Array<String>} [options.appliedTags] The IDs of the set of tags that have been applied to a thread in a forum channel (threads created in forum channels only, max 5)
+     * @arg {Array<String>} [options.appliedTags] The IDs of the set of tags that have been applied to a thread in a forum/media channel (threads created in forum/media channels only, max 5)
      * @arg {Boolean} [options.archived] The archive status of the channel (thread channels only)
      * @arg {Number} [options.autoArchiveDuration] The duration in minutes to automatically archive the thread after recent activity, either 60, 1440, 4320 or 10080 (thread channels only)
-     * @arg {Array<Object>} [options.availableTags] The available tags that can be applied to threads in a forum channel. See [the official Discord API documentation entry](https://discord.com/developers/docs/resources/channel#forum-tag-object) for object structure (forum channels only, max 20)
+     * @arg {Array<Object>} [options.availableTags] The available tags that can be applied to threads in a forum/media channel. See [the official Discord API documentation entry](https://discord.com/developers/docs/resources/channel#forum-tag-object) for object structure (forum/media channels only, max 20)
      * @arg {Number} [options.bitrate] The bitrate of the channel (guild voice channels only)
-     * @arg {Number} [options.defaultAutoArchiveDuration] The default duration of newly created threads in minutes to automatically archive the thread after inactivity (60, 1440, 4320, 10080) (text/news/forum channels only)
+     * @arg {Number} [options.defaultAutoArchiveDuration] The default duration of newly created threads in minutes to automatically archive the thread after inactivity (60, 1440, 4320, 10080) (text/news/forum/media channels only)
      * @arg {Number} [options.defaultForumLayout] The default forum layout type used to display posts in forum channels (forum channels only)
-     * @arg {Object} [options.defaultReactionEmoji] The emoji to show in the add reaction button on a thread in a forum channel (forum channels only)
-     * @arg {Number} [options.defaultSortOrder] The default sort order type used to order posts in forum channels (forum channels only)
-     * @arg {Number} [options.defaultThreadRateLimitPerUser] The initial rateLimitPerUser to set on newly created threads in a channel (text/forum channels only)
-     * @arg {Number} [options.flags] The flags for the channel combined as a bitfield (thread/forum channels only). Note: `PINNED` can only be set for threads in forum channels, and `REQUIRE_TAG` can only be set for forum channels
+     * @arg {Object} [options.defaultReactionEmoji] The emoji to show in the add reaction button on a thread in a forum/media channel (forum/media channels only)
+     * @arg {Number} [options.defaultSortOrder] The default sort order type used to order posts in forum/media channels (forum/media channels only)
+     * @arg {Number} [options.defaultThreadRateLimitPerUser] The initial rateLimitPerUser to set on newly created threads in a channel (text/forum/media channels only)
+     * @arg {Number} [options.flags] The flags for the channel combined as a bitfield (thread/forum/media channels only). Note: `PINNED` can only be set for threads in forum/media channels, `REQUIRE_TAG` can only be set for forum/media channels, and `HIDE_MEDIA_DOWNLOAD_OPTIONS` can only be set for media channels
      * @arg {String} [options.icon] The icon of the channel as a base64 data URI (group channels only). Note: base64 strings alone are not base64 data URI strings
      * @arg {Boolean} [options.invitable] Whether non-moderators can add other non-moderators to the channel (private thread channels only)
      * @arg {Boolean} [options.locked] The lock status of the channel (thread channels only)
@@ -1556,7 +1556,7 @@ class Client extends EventEmitter {
      * @arg {String?} [options.parentID] The ID of the parent channel category for this channel (guild text/voice channels only)
      * @arg {Array<Object>} [options.permissionOverwrites] An array containing permission overwrite objects
      * @arg {Number} [options.position] The sorting position of the channel (guild channels only)
-     * @arg {Number} [options.rateLimitPerUser] The time in seconds a user has to wait before sending another message (0-21600) (does not affect bots or users with manageMessages/manageChannel permissions) (text/voice/stage/forum channels only)
+     * @arg {Number} [options.rateLimitPerUser] The time in seconds a user has to wait before sending another message (0-21600) (does not affect bots or users with manageMessages/manageChannel permissions) (text/voice/stage/forum/media channels only)
      * @arg {String?} [options.rtcRegion] The RTC region ID of the channel (automatic if `null`) (guild voice channels only)
      * @arg {String} [options.topic] The topic of the channel (guild text channels only)
      * @arg {Number} [options.userLimit] The channel user limit (guild voice channels only)
