+++ lib/Constants.js
@@ -165,8 +165,9 @@ module.exports.ButtonStyles = {
 };
 
 module.exports.ChannelFlags = {
-    PINNED:      1 << 1,
-    REQUIRE_TAG: 1 << 4
+    PINNED:                      1 << 1,
+    REQUIRE_TAG:                 1 << 4,
+    HIDE_MEDIA_DOWNLOAD_OPTIONS: 1 << 15
 };
 
 module.exports.ChannelTypes = {
@@ -182,7 +183,8 @@ module.exports.ChannelTypes = {
     GUILD_PRIVATE_THREAD: 12,
     GUILD_STAGE_VOICE:    13, GUILD_STAGE: 13, // [DEPRECATED]
 
-    GUILD_FORUM: 15
+    GUILD_FORUM: 15,
+    GUILD_MEDIA: 16
 };
 
 module.exports.ComponentTypes = {
