+++ lib/gateway/Shard.js
@@ -1769,20 +1769,20 @@ class Shard extends EventEmitter {
                  * @event Client#channelUpdate
                  * @prop {CategoryChannel | DMChannel | ForumChannel | GuildChannel | TextChannel | VoiceChannel | NewsChannel} channel The updated channel
                  * @prop {Object} oldChannel The old channel data
-                 * @prop {Array<Object>} oldChannel.availableTags The available tags that can be applied to threads in a forum channel. See [the official Discord API documentation entry](https://discord.com/developers/docs/resources/channel#forum-tag-object) for object structure (forum channels only, max 20)
+                 * @prop {Array<Object>} oldChannel.availableTags The available tags that can be applied to threads in a forum/media channel. See [the official Discord API documentation entry](https://discord.com/developers/docs/resources/channel#forum-tag-object) for object structure (forum/media channels only, max 20)
                  * @prop {Number} oldChannel.bitrate The bitrate of the channel (voice channels only)
-                 * @prop {Number} oldChannel.defaultAutoArchiveDuration The default duration of newly created threads in minutes to automatically archive the thread after inactivity (60, 1440, 4320, 10080) (text/news/forum channels only)
+                 * @prop {Number} oldChannel.defaultAutoArchiveDuration The default duration of newly created threads in minutes to automatically archive the thread after inactivity (60, 1440, 4320, 10080) (text/news/forum/media channels only)
                  * @prop {Number} oldChannel.defaultForumLayout The default forum layout type used to display posts in forum channels (forum channels only)
-                 * @prop {Object} oldChannel.defaultReactionEmoji The emoji to show in the add reaction button on a thread in a forum channel (forum channels only)
-                 * @prop {Number} oldChannel.defaultSortOrder The default sort order type used to order posts in forum channels (forum channels only)
-                 * @prop {Number} oldChannel.defaultThreadRateLimitPerUser The initial rateLimitPerUser to set on newly created threads in a channel (text/forum channels only)
-                 * @prop {Number?} oldChannel.flags The flags for the channel combined as a bitfield (thread/forum channels only)
+                 * @prop {Object} oldChannel.defaultReactionEmoji The emoji to show in the add reaction button on a thread in a forum/media channel (forum/media channels only)
+                 * @prop {Number} oldChannel.defaultSortOrder The default sort order type used to order posts in forum/media channels (forum/media channels only)
+                 * @prop {Number} oldChannel.defaultThreadRateLimitPerUser The initial rateLimitPerUser to set on newly created threads in a channel (text/forum/media channels only)
+                 * @prop {Number?} oldChannel.flags The flags for the channel combined as a bitfield (thread/forum/media channels only)
                  * @prop {String} oldChannel.name The name of the channel
                  * @prop {Boolean} oldChannel.nsfw Whether the channel is NSFW or not (text channels only)
                  * @prop {String?} oldChannel.parentID The ID of the category this channel belongs to (guild channels only)
                  * @prop {Collection} oldChannel.permissionOverwrites Collection of PermissionOverwrites in this channel (guild channels only)
                  * @prop {Number} oldChannel.position The position of the channel (guild channels only)
-                 * @prop {Number?} oldChannel.rateLimitPerUser The time in seconds a user has to wait before sending another message (0-21600) (text/voice/stage/forum channels only)
+                 * @prop {Number?} oldChannel.rateLimitPerUser The time in seconds a user has to wait before sending another message (0-21600) (text/voice/stage/forum/media channels only)
                  * @prop {String?} oldChannel.rtcRegion The RTC region ID of the channel (automatic when `null`) (voice channels only)
                  * @prop {String?} oldChannel.topic The topic of the channel (text channels only)
                  * @prop {Number} oldChannel.type The type of the old channel (text/news channels only)
@@ -2371,7 +2371,7 @@ class Shard extends EventEmitter {
                  * @event Client#threadUpdate
                  * @prop {NewsThreadChannel | PrivateThreadChannel | PublicThreadChannel} channel The updated channel
                  * @prop {Object?} oldChannel The old thread channel. This will be null if the channel was uncached
-                 * @prop {Array<String>?} oldChannel.appliedTags The IDs of the set of tags that have been applied to a thread in a forum channel
+                 * @prop {Array<String>?} oldChannel.appliedTags The IDs of the set of tags that have been applied to a thread in a forum/media channel
                  * @prop {Number} oldChannel.autoArchiveDuration The duration in minutes to automatically archive the thread after recent activity, either 60, 1440, 4320 or 10080
                  * @prop {Number} oldChannel.flags The flags for the channel combined as a bitfield
                  * @prop {String} oldChannel.name The name of the channel
