+++ lib/structures/Channel.js
@@ -56,6 +56,9 @@ class Channel extends Base {
             case ChannelTypes.GUILD_FORUM: {
                 return new ForumChannel(data, client);
             }
+            case ChannelTypes.GUILD_MEDIA: {
+                return new MediaChannel(data, client);
+            }
         }
         if(data.guild_id) {
             if(data.last_message_id !== undefined) {
@@ -85,6 +88,7 @@ const DMChannel = require("./DMChannel");
 const GuildChannel = require("./GuildChannel");
 const ForumChannel = require("./ForumChannel");
 const GroupChannel = require("./GroupChannel");
+const MediaChannel = require("./MediaChannel");
 const NewsChannel = require("./NewsChannel");
 const NewsThreadChannel = require("./NewsThreadChannel");
 const PrivateThreadChannel = require("./PrivateThreadChannel");
