+++ lib/structures/GuildChannel.js
@@ -7,7 +7,7 @@ const {Permissions} = require("../Constants");
 /**
  * Represents a guild channel. You also probably want to look at CategoryChannel, ForumChannel, NewsChannel, StageChannel, TextChannel, ThreadChannel, and VoiceChannel. See Channel for more properties
  * @extends Channel
- * @prop {Number} flags The channel's flags (see constants). Currently only applies to thread/forum channels
+ * @prop {Number} flags The channel's flags (see constants). Currently only applies to thread/forum/media channels
  * @prop {Guild} guild The guild that owns the channel
  * @prop {String} name The name of the channel
  * @prop {String?} parentID The ID of the category (or channel for threads) this channel belongs to
