+++ lib/structures/PublicThreadChannel.js
@@ -5,7 +5,7 @@ const ThreadChannel = require("./ThreadChannel");
 /**
  * Represents a public thread channel. See ThreadChannel for more properties
  * @extends ThreadChannel
- * @prop {Array<String>?} appliedTags The IDs of the set of tags that have been applied to a thread in a forum channel (threads created in forum channels only, max 5)
+ * @prop {Array<String>?} appliedTags The IDs of the set of tags that have been applied to a thread in a forum/media channel (threads created in forum/media channels only, max 5)
  */
 class PublicThreadChannel extends ThreadChannel {
     update(data) {
